refactor: split Node class into more mixins
